rename: 디렉터리명 복수형으로 수정
